+++ api/models/Group.ts
@@ -27,39 +27,10 @@ import {
   tryToMatchRecordingToStation,
 } from "@api/V1/recordingUtil";
 import { Device } from "./Device";
-import { GroupId, UserId, LatLng } from "@typedefs/api/common";
+import { GroupId, UserId } from "@typedefs/api/common";
 import { ApiGroupSettings } from "@typedefs/api/group";
+import { locationsAreEqual } from "./util/util";
 
-const EPSILON = 0.000000000001;
-
-export const canonicalLatLng = (
-  location: LatLng | { coordinates: [number, number] } | [number, number]
-): LatLng => {
-  if (Array.isArray(location)) {
-    return { lat: location[0], lng: location[1] };
-  } else if (location.hasOwnProperty("coordinates")) {
-    // Lat lng is stored in the database as lng/lat (X,Y).
-    // If we get lat/lng in this format we are getting it from the DB.
-    return {
-      lat: (location as { coordinates: [number, number] }).coordinates[1],
-      lng: (location as { coordinates: [number, number] }).coordinates[0],
-    };
-  }
-  return location as LatLng;
-};
-
-export const locationsAreEqual = (
-  a: LatLng | { coordinates: [number, number] },
-  b: LatLng | { coordinates: [number, number] }
-): boolean => {
-  const canonicalA = canonicalLatLng(a);
-  const canonicalB = canonicalLatLng(b);
-  // NOTE: We need to compare these numbers with an epsilon value, otherwise we get floating-point precision issues.
-  return (
-    Math.abs(canonicalA.lat - canonicalB.lat) < EPSILON &&
-    Math.abs(canonicalA.lng - canonicalB.lng) < EPSILON
-  );
-};
 
 export const stationLocationHasChanged = (
   oldStation: Station,
